+++ tests/drivers/flash/common/testcase.yaml
@@ -65,7 +65,7 @@ tests:
   drivers.flash.common.mr_canhubk3_sfdp_runtime:
     platform_allow: mr_canhubk3
     extra_configs:
-      - CONFIG_FLASH_NXP_S32_QSPI_NOR_SFDP_RUNTIME=y
+      - CONFIG_FLASH_NXP_S32_QSPI_SFDP_RUNTIME=y
   drivers.flash.common.spi_nor:
     platform_allow: nrf52840dk/nrf52840
     extra_configs:
